+++ packages/article/__tests__/android/article.android.test.js
@@ -1,0 +1,7 @@
+import shared from "../shared";
+import nativeShared from "../shared-native";
+
+describe("Article tests on android", () => {
+  shared();
+  nativeShared();
+});