+++ packages/article/src/__tests__/ios/article.ios.test.js
@@ -1,5 +1,0 @@
-import shared from "../shared";
-
-describe("Article test on ios", () => {
-  shared();
-});