+++ packages/article/src/__tests__/web/jest.config.js
@@ -1,6 +1,0 @@
-const jestConfigurator = require("@times-components/jest-configurator").default;
-
-module.exports = jestConfigurator("web", __dirname, [
-  "data-helper.js",
-  "article-content.js"
-]);