+++ packages/article/src/article-tracking-context.js
@@ -4,7 +4,7 @@ import { withTrackingContext } from "@times-components/tracking";
 export default Component =>
   withTrackingContext(Component, {
     trackingObjectName: "Article",
-    getAttrs: ({ article } = {}) => ({
+    getAttrs: ({ article }) => ({
       byline: get(article, "byline[0].children[0].attributes.value", ""),
       headline: get(article, "headline", ""),
       publishedTime: get(article, "publishedTime", ""),