+++ packages/link/src/__tests__/ios/link.ios.test.js
@@ -1,7 +1,0 @@
-import { Text } from "react-native";
-import test from "../shared";
-import Link, { TextLink } from "../../link";
-
-describe("Link test on ios", () => {
-  test(Link, TextLink, Text);
-});