+++ packages/slice/src/templates/opinionandtwo/config.web.js
@@ -39,7 +39,7 @@ export const getConfigWrapper = ({ supportCount }) => {
       }
       .opinionImageContainerClass {
         bottom: 0;
-        margin-bottom: ${spacing(-2)};
+        margin-bottom: ${spacing(1)};
         min-width: 115px;
         position: absolute;
         right: 0;
@@ -93,6 +93,7 @@ export const getConfigWrapper = ({ supportCount }) => {
           min-width: 165px;
         `;
         return `
+          margin-bottom: ${spacing(2)};
           min-width: 152px;
         `;
       };