+++ packages/styleguide/__tests__/shared.js
@@ -1,0 +1,18 @@
+import { Text } from "react-native";
+import React from "react";
+import renderer from "react-test-renderer";
+import { Animations } from "../src/styleguide";
+
+export default () => {
+  it("should render a FadeIn animation wrapper component", () => {
+    const tree = renderer
+      .create(
+        <Animations.FadeIn>
+          <Text>Hello World</Text>
+        </Animations.FadeIn>
+      )
+      .toJSON();
+
+    expect(tree).toMatchSnapshot();
+  });
+};