+++ packages/video-label/src/__tests__/beautify-title-without-spacing.js
@@ -1,7 +1,0 @@
-import beautifyTitle from "../beautify-title";
-
-export default () => {
-  it("does capitalise every letter", () => {
-    expect(beautifyTitle("title")).toMatch(/TITLE/);
-  });
-};