+++ packages/author-head/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.22.23"></a>
+## [0.22.23](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.22...@times-components/author-head@0.22.23) (2018-04-11)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.22.22"></a>
+## [0.22.22](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.19...@times-components/author-head@0.22.22) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.22.21"></a>
+## [0.22.21](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.19...@times-components/author-head@0.22.21) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.22.20"></a>
+## [0.22.20](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.19...@times-components/author-head@0.22.20) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.22.19"></a>
+## [0.22.19](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.18...@times-components/author-head@0.22.19) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.22.18"></a>
+## [0.22.18](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.17...@times-components/author-head@0.22.18) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.22.17"></a>
+## [0.22.17](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.16...@times-components/author-head@0.22.17) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.22.16"></a>
+## [0.22.16](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.15...@times-components/author-head@0.22.16) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
 <a name="0.22.15"></a>
 ## [0.22.15](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.14...@times-components/author-head@0.22.15) (2018-04-06)