+++ packages/card/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.25.16"></a>
+## [0.25.16](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.15...@times-components/card@0.25.16) (2018-04-11)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
+<a name="0.25.15"></a>
+## [0.25.15](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.12...@times-components/card@0.25.15) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
+<a name="0.25.14"></a>
+## [0.25.14](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.12...@times-components/card@0.25.14) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
+<a name="0.25.13"></a>
+## [0.25.13](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.12...@times-components/card@0.25.13) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
+<a name="0.25.12"></a>
+## [0.25.12](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.11...@times-components/card@0.25.12) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
+<a name="0.25.11"></a>
+## [0.25.11](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.9...@times-components/card@0.25.11) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
+<a name="0.25.10"></a>
+## [0.25.10](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.9...@times-components/card@0.25.10) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
+<a name="0.25.9"></a>
+## [0.25.9](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.8...@times-components/card@0.25.9) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
 <a name="0.25.8"></a>
 ## [0.25.8](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.7...@times-components/card@0.25.8) (2018-04-06)