+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.1"></a>
+## [0.6.1](https://github.com/newsuk/times-components/compare/@times-components/icons@0.6.0...@times-components/icons@0.6.1) (2018-04-11)
+
+
+
+
+**Note:** Version bump only for package @times-components/icons
+
+<a name="0.6.0"></a>
+# [0.6.0](https://github.com/newsuk/times-components/compare/@times-components/icons@0.3.19...@times-components/icons@0.6.0) (2018-04-10)
+
+
+### Features
+
+* rewrite the story ([#825](https://github.com/newsuk/times-components/issues/825)) ([c08b823](https://github.com/newsuk/times-components/commit/c08b823))
+
+
+
+
+<a name="0.5.0"></a>
+# [0.5.0](https://github.com/newsuk/times-components/compare/@times-components/icons@0.3.19...@times-components/icons@0.5.0) (2018-04-10)
+
+
+### Features
+
+* rewrite the story ([#825](https://github.com/newsuk/times-components/issues/825)) ([c08b823](https://github.com/newsuk/times-components/commit/c08b823))
+
+
+
+
+<a name="0.4.0"></a>
+# [0.4.0](https://github.com/newsuk/times-components/compare/@times-components/icons@0.3.19...@times-components/icons@0.4.0) (2018-04-10)
+
+
+### Features
+
+* rewrite the story ([#825](https://github.com/newsuk/times-components/issues/825)) ([c08b823](https://github.com/newsuk/times-components/commit/c08b823))
+
+
+
+
+<a name="0.3.19"></a>
+## [0.3.19](https://github.com/newsuk/times-components/compare/@times-components/icons@0.3.18...@times-components/icons@0.3.19) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/icons
+
 <a name="0.3.18"></a>
 ## [0.3.18](https://github.com/newsuk/times-components/compare/@times-components/icons@0.3.17...@times-components/icons@0.3.18) (2018-04-06)