+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,98 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.0.14"></a>
+## [0.0.14](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.13...@times-components/related-articles@0.0.14) (2018-04-11)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.13"></a>
+## [0.0.13](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.12...@times-components/related-articles@0.0.13) (2018-04-11)
+
+
+### Bug Fixes
+
+* updated path to cutout image used in opinion slice ([#835](https://github.com/newsuk/times-components/issues/835)) ([e1db56b](https://github.com/newsuk/times-components/commit/e1db56b))
+
+
+
+
+<a name="0.0.12"></a>
+## [0.0.12](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.9...@times-components/related-articles@0.0.12) (2018-04-10)
+
+
+### Bug Fixes
+
+* fix opinion layout ([#839](https://github.com/newsuk/times-components/issues/839)) ([e54f403](https://github.com/newsuk/times-components/commit/e54f403))
+
+
+
+
+<a name="0.0.11"></a>
+## [0.0.11](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.9...@times-components/related-articles@0.0.11) (2018-04-10)
+
+
+### Bug Fixes
+
+* fix opinion layout ([#839](https://github.com/newsuk/times-components/issues/839)) ([e54f403](https://github.com/newsuk/times-components/commit/e54f403))
+
+
+
+
+<a name="0.0.10"></a>
+## [0.0.10](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.9...@times-components/related-articles@0.0.10) (2018-04-10)
+
+
+### Bug Fixes
+
+* fix opinion layout ([#839](https://github.com/newsuk/times-components/issues/839)) ([e54f403](https://github.com/newsuk/times-components/commit/e54f403))
+
+
+
+
+<a name="0.0.9"></a>
+## [0.0.9](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.8...@times-components/related-articles@0.0.9) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.8"></a>
+## [0.0.8](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.7...@times-components/related-articles@0.0.8) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.7"></a>
+## [0.0.7](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.5...@times-components/related-articles@0.0.7) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.6"></a>
+## [0.0.6](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.5...@times-components/related-articles@0.0.6) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.5"></a>
+## [0.0.5](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.4...@times-components/related-articles@0.0.5) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
 <a name="0.0.4"></a>
 ## [0.0.4](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.3...@times-components/related-articles@0.0.4) (2018-04-06)