+++ packages/slice/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.17.11"></a>
+## [0.17.11](https://github.com/newsuk/times-components/compare/@times-components/slice@0.17.10...@times-components/slice@0.17.11) (2018-04-11)
+
+
+
+
+**Note:** Version bump only for package @times-components/slice
+
+<a name="0.17.10"></a>
+## [0.17.10](https://github.com/newsuk/times-components/compare/@times-components/slice@0.17.7...@times-components/slice@0.17.10) (2018-04-10)
+
+
+### Bug Fixes
+
+* fix opinion layout ([#839](https://github.com/newsuk/times-components/issues/839)) ([e54f403](https://github.com/newsuk/times-components/commit/e54f403))
+
+
+
+
+<a name="0.17.9"></a>
+## [0.17.9](https://github.com/newsuk/times-components/compare/@times-components/slice@0.17.7...@times-components/slice@0.17.9) (2018-04-10)
+
+
+### Bug Fixes
+
+* fix opinion layout ([#839](https://github.com/newsuk/times-components/issues/839)) ([e54f403](https://github.com/newsuk/times-components/commit/e54f403))
+
+
+
+
+<a name="0.17.8"></a>
+## [0.17.8](https://github.com/newsuk/times-components/compare/@times-components/slice@0.17.7...@times-components/slice@0.17.8) (2018-04-10)
+
+
+### Bug Fixes
+
+* fix opinion layout ([#839](https://github.com/newsuk/times-components/issues/839)) ([e54f403](https://github.com/newsuk/times-components/commit/e54f403))
+
+
+
+
+<a name="0.17.7"></a>
+## [0.17.7](https://github.com/newsuk/times-components/compare/@times-components/slice@0.17.6...@times-components/slice@0.17.7) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/slice
+
+<a name="0.17.6"></a>
+## [0.17.6](https://github.com/newsuk/times-components/compare/@times-components/slice@0.17.5...@times-components/slice@0.17.6) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/slice
+
 <a name="0.17.5"></a>
 ## [0.17.5](https://github.com/newsuk/times-components/compare/@times-components/slice@0.17.4...@times-components/slice@0.17.5) (2018-04-06)