+++ packages/video-label/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.27"></a>
+## [0.1.27](https://github.com/newsuk/times-components/compare/@times-components/video-label@0.1.26...@times-components/video-label@0.1.27) (2018-04-11)
+
+
+
+
+**Note:** Version bump only for package @times-components/video-label
+
+<a name="0.1.26"></a>
+## [0.1.26](https://github.com/newsuk/times-components/compare/@times-components/video-label@0.1.23...@times-components/video-label@0.1.26) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/video-label
+
+<a name="0.1.25"></a>
+## [0.1.25](https://github.com/newsuk/times-components/compare/@times-components/video-label@0.1.23...@times-components/video-label@0.1.25) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/video-label
+
+<a name="0.1.24"></a>
+## [0.1.24](https://github.com/newsuk/times-components/compare/@times-components/video-label@0.1.23...@times-components/video-label@0.1.24) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/video-label
+
+<a name="0.1.23"></a>
+## [0.1.23](https://github.com/newsuk/times-components/compare/@times-components/video-label@0.1.22...@times-components/video-label@0.1.23) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/video-label
+
 <a name="0.1.22"></a>
 ## [0.1.22](https://github.com/newsuk/times-components/compare/@times-components/video-label@0.1.21...@times-components/video-label@0.1.22) (2018-04-06)