+++ packages/watermark/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.4.22"></a>
+## [0.4.22](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.4.21...@times-components/watermark@0.4.22) (2018-04-11)
+
+
+
+
+**Note:** Version bump only for package @times-components/watermark
+
+<a name="0.4.21"></a>
+## [0.4.21](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.4.18...@times-components/watermark@0.4.21) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/watermark
+
+<a name="0.4.20"></a>
+## [0.4.20](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.4.18...@times-components/watermark@0.4.20) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/watermark
+
+<a name="0.4.19"></a>
+## [0.4.19](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.4.18...@times-components/watermark@0.4.19) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/watermark
+
+<a name="0.4.18"></a>
+## [0.4.18](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.4.17...@times-components/watermark@0.4.18) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/watermark
+
 <a name="0.4.17"></a>
 ## [0.4.17](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.4.16...@times-components/watermark@0.4.17) (2018-04-06)